requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
ichain_configurator==0.5.0
ichain-logger==0.5.1
requests==2.32.3
backoff==2.2.1
pandas== 2.2.3
numpy==2.2.2
pydantic==2.10.6
email_validator==2.2.0
PyJWT==2.10.1